Day17 : Schedule Object

+++ src/main/java/com/ezen/antpeople/repository/ScheRepository.java
@@ -9,7 +9,7 @@ import com.ezen.antpeople.entity.ScheEntity;
 import com.ezen.antpeople.entity.UserEntity;
 
 @Repository("schedule")
-public interface ScheRepository extends JpaRepository<ScheEntity, Long> {
+public interface ScheRepository extends JpaRepository<ScheEntity, Integer> {
 	List<ScheEntity> findByFromUser(UserEntity user);
 	List<ScheEntity> findByFromUserAndStartDate(UserEntity user, String startDate);
